requirements.txt
numpy==1.22.0
falcon==3.0.1
falcon-marshmallow==0.3.0
marshmallow==3.13.0
marshmallow-union
diffusers==0.20.2
torch==2.0.1
gunicorn==21.2.0
pyjwt==1.7.1
pika==0.13.1
pymongo==4.5.0
